pyproject.toml
[project]
name = "figma-mcp-server"
version = "1.0.0"
description = "MCP-сервер для работы с Figma API и дизайн-системой"
authors = [{name = "Your Name"}]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastmcp>=2.10.0",
    "aiohttp>=3.9.0",
    "aiofiles>=23.2.0",
    "python-dotenv>=1.0.0",
    "prometheus-client>=0.19.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
[tool.hatch.build.targets.wheel]
packages = ["src"]